+++ Glass/Components/SlidingMessageFrame.lua
@@ -3,17 +3,20 @@ local TP = Core:GetModule("TextProcessing")
 
 local AceHook = Core.Libs.AceHook
 
+local LibEasing = Core.Libs.LibEasing
 local lodash = Core.Libs.lodash
 local drop, reduce, take = lodash.drop, lodash.reduce, lodash.take
 
+local CreateFadingFrame = Core.Components.CreateFadingFrame
 local CreateMessageLinePool = Core.Components.CreateMessageLinePool
+
+local Colors = Constants.COLORS
 
 local MOUSE_ENTER = Constants.EVENTS.MOUSE_ENTER
 local MOUSE_LEAVE = Constants.EVENTS.MOUSE_LEAVE
 local UPDATE_CONFIG = Constants.EVENTS.UPDATE_CONFIG
 
 -- luacheck: push ignore 113
-local C_Timer = C_Timer
 local CreateFrame = CreateFrame
 local CreateObjectPool = CreateObjectPool
 local Mixin = Mixin
@@ -38,6 +41,8 @@ function SlidingMessageFrameMixin:Init(chatFrame)
     incomingMessages = {},
     messages = {},
     isCombatLog = false,
+    scrollAtBottom = true,
+    unreadMessages = false,
   }
   self.chatFrame = chatFrame
 
@@ -67,16 +72,157 @@ function SlidingMessageFrameMixin:Init(chatFrame)
   -- Chat scroll frame
   self:SetHeight(self.config.height + self.config.overflowHeight)
   self:SetWidth(self.config.width)
-  self:SetPoint("BOTTOMLEFT", 0, self.config.overflowHeight * -1)
+  self:SetPoint("TOPLEFT", 0, (Constants.DOCK_HEIGHT + 5) * -1)
 
   -- Set initial scroll position
   self:SetVerticalScroll(self.config.overflowHeight)
 
-  if self.bg == nil then
-    self.bg = self:CreateTexture(nil, "BACKGROUND")
+  -- Overlay
+  if self.overlay == nil then
+    local overlayOpacity = 0.65
+
+    self.overlay = CreateFadingFrame("Frame", nil, self)
+    self.overlay:SetHeight(64)
+    self.overlay:SetPoint("TOPLEFT", 0, (self.config.height - 62) * -1)
+    self.overlay:SetPoint("TOPRIGHT", 0, (self.config.height - 62) * -1)
+    self.overlay:SetFadeInDuration(0.3)
+    self.overlay:SetFadeOutDuration(0.15)
+    self.overlay:QuickHide()
+
+    self.overlay.mask = self.overlay:CreateMaskTexture()
+    self.overlay.mask:SetTexture("Interface\\Addons\\Glass\\Assets\\overlayMask", "CLAMP", "CLAMPTOBLACKADDITIVE")
+    self.overlay.mask:SetSize(16, 64)
+    self.overlay.mask:SetPoint("CENTER", 0, -32)
+
+    self.overlay.leftBg = self.overlay:CreateTexture(nil, "BACKGROUND")
+    self.overlay.leftBg:SetPoint("TOPLEFT")
+    self.overlay.leftBg:SetPoint("BOTTOMLEFT")
+    self.overlay.leftBg:SetWidth(15)
+    self.overlay.leftBg:SetColorTexture(1, 1, 1, 1)
+    self.overlay.leftBg:SetGradientAlpha(
+      "HORIZONTAL",
+      Colors.codGray.r, Colors.codGray.g, Colors.codGray.b, 0,
+      Colors.codGray.r, Colors.codGray.g, Colors.codGray.b, overlayOpacity
+    )
+    self.overlay.leftBg:AddMaskTexture(self.overlay.mask)
+
+    self.overlay.centerBg = self.overlay:CreateTexture(nil, "BACKGROUND")
+    self.overlay.centerBg:SetPoint("TOPLEFT", 15, 0)
+    self.overlay.centerBg:SetPoint("BOTTOMRIGHT", -15, 0)
+    self.overlay.centerBg:SetColorTexture(
+      Colors.codGray.r,
+      Colors.codGray.g,
+      Colors.codGray.b,
+      overlayOpacity
+    )
+    self.overlay.centerBg:AddMaskTexture(self.overlay.mask)
+
+    self.overlay.rightBg = self.overlay:CreateTexture(nil, "BACKGROUND")
+    self.overlay.rightBg:SetPoint("TOPRIGHT")
+    self.overlay.rightBg:SetPoint("BOTTOMRIGHT")
+    self.overlay.rightBg:SetWidth(15)
+    self.overlay.rightBg:SetColorTexture(1, 1, 1, 1)
+    self.overlay.rightBg:SetGradientAlpha(
+      "HORIZONTAL",
+      Colors.codGray.r, Colors.codGray.g, Colors.codGray.b, overlayOpacity,
+      Colors.codGray.r, Colors.codGray.g, Colors.codGray.b, 0
+    )
+    self.overlay.rightBg:AddMaskTexture(self.overlay.mask)
+
+    -- Down arrow icon
+    self.overlay.icon = self.overlay:CreateTexture(nil, "ARTWORK")
+    self.overlay.icon:SetTexture("Interface\\Addons\\Glass\\Glass\\Assets\\snapToBottomIcon")
+    self.overlay.icon:SetSize(16, 16)
+    self.overlay.icon:SetPoint("BOTTOMLEFT", 15, 5)
+
+    -- See new messages click area
+    self.overlay.snapToBottomFrame = CreateFrame("Frame", nil, self.overlay)
+    self.overlay.snapToBottomFrame:SetHeight(20)
+    self.overlay.snapToBottomFrame:SetPoint("BOTTOMLEFT")
+    self.overlay.snapToBottomFrame:SetPoint("BOTTOMRIGHT")
+
+    -- Snap to bottom on click
+    self.overlay.snapToBottomFrame:SetScript("OnMouseDown", function ()
+      self.state.scrollAtBottom = true
+      self.state.unreadMessages = false
+      self.overlay:Hide()
+      self.overlay.newMessageHighlightFrame:Hide()
+
+      local startOffset = math.max(
+        self:GetVerticalScrollRange() - self.config.height * 2,
+        self:GetVerticalScroll()
+      )
+      local endOffset = self:GetVerticalScrollRange()
+
+      LibEasing:Ease(
+        function (offset) self:SetVerticalScroll(offset) end,
+        startOffset,
+        endOffset,
+        0.3,
+        LibEasing.OutCubic,
+        function ()
+          self:SetHeight(self.config.height + self.config.overflowHeight)
+        end
+      )
+    end)
+
+    -- See new messages frame
+    self.overlay.newMessageHighlightFrame = CreateFadingFrame("Frame", nil, self.overlay)
+    self.overlay.newMessageHighlightFrame:SetHeight(20)
+    self.overlay.newMessageHighlightFrame:SetPoint("BOTTOMLEFT")
+    self.overlay.newMessageHighlightFrame:SetPoint("BOTTOMRIGHT")
+    self.overlay.newMessageHighlightFrame:SetFadeInDuration(0.15)
+    self.overlay.newMessageHighlightFrame:SetFadeOutDuration(0.15)
+
+    -- New messages text
+    self.overlay.newMessageHighlightFrame.text = self.overlay.newMessageHighlightFrame:CreateFontString(
+      nil, "ARTWORK", "GlassMessageFont"
+    )
+    self.overlay.newMessageHighlightFrame.text:SetTextColor(Colors.apache.r, Colors.apache.g, Colors.apache.b)
+    self.overlay.newMessageHighlightFrame.text:SetPoint("BOTTOMLEFT", 30, 10)
+    self.overlay.newMessageHighlightFrame.text:SetText("Unread messages")
+
+    -- Highlight line
+    self.overlay.newMessageHighlightFrame.leftBg = self.overlay.newMessageHighlightFrame:CreateTexture(
+      nil, "BACKGROUND"
+    )
+    self.overlay.newMessageHighlightFrame.leftBg:SetPoint("BOTTOMLEFT")
+    self.overlay.newMessageHighlightFrame.leftBg:SetHeight(1)
+    self.overlay.newMessageHighlightFrame.leftBg:SetWidth(15)
+    self.overlay.newMessageHighlightFrame.leftBg:SetColorTexture(1, 1, 1, 1)
+    self.overlay.newMessageHighlightFrame.leftBg:SetGradientAlpha(
+      "HORIZONTAL",
+      Colors.apache.r, Colors.apache.g, Colors.apache.b, 0,
+      Colors.apache.r, Colors.apache.g, Colors.apache.b, overlayOpacity
+    )
+
+    self.overlay.newMessageHighlightFrame.centerBg = self.overlay.newMessageHighlightFrame:CreateTexture(
+      nil, "BACKGROUND"
+    )
+    self.overlay.newMessageHighlightFrame.centerBg:SetHeight(1)
+    self.overlay.newMessageHighlightFrame.centerBg:SetPoint("BOTTOMLEFT", 15, 0)
+    self.overlay.newMessageHighlightFrame.centerBg:SetPoint("BOTTOMRIGHT", -15, 0)
+    self.overlay.newMessageHighlightFrame.centerBg:SetColorTexture(
+      Colors.apache.r,
+      Colors.apache.g,
+      Colors.apache.b,
+      overlayOpacity
+    )
+
+    self.overlay.newMessageHighlightFrame.rightBg = self.overlay.newMessageHighlightFrame:CreateTexture(
+      nil, "BACKGROUND"
+    )
+    self.overlay.newMessageHighlightFrame.rightBg:SetPoint("BOTTOMRIGHT")
+    self.overlay.newMessageHighlightFrame.rightBg:SetHeight(1)
+    self.overlay.newMessageHighlightFrame.rightBg:SetWidth(15)
+    self.overlay.newMessageHighlightFrame.rightBg:SetColorTexture(1, 1, 1, 1)
+    self.overlay.newMessageHighlightFrame.rightBg:SetGradientAlpha(
+      "HORIZONTAL",
+      Colors.apache.r, Colors.apache.g, Colors.apache.b, overlayOpacity,
+      Colors.apache.r, Colors.apache.g, Colors.apache.b, 0
+    )
+    self.overlay.newMessageHighlightFrame:QuickHide()
   end
-  self.bg:SetAllPoints()
-  self.bg:SetColorTexture(0, 1, 0, 0)
 
   self.timeElapsed = 0
   self:SetScript("OnUpdate", function (frame, elapsed)
@@ -85,21 +231,45 @@ function SlidingMessageFrameMixin:Init(chatFrame)
 
   -- Scrolling
   self:SetScript("OnMouseWheel", function (frame, delta)
-    local currentScrollOffset = self:GetVerticalScroll()
-    local scrollRange = self:GetVerticalScrollRange()
+    local maxScroll = (
+      self.state.scrollAtBottom and
+      self:GetVerticalScrollRange() + self.config.overflowHeight
+      or self:GetVerticalScrollRange()
+    )
+    local minScroll = self.config.height + self.config.overflowHeight
+    local scrollValue
 
-    -- Adjust scroll
-    if delta < 0 and currentScrollOffset < scrollRange + self.config.overflowHeight then
-      self:SetVerticalScroll(math.min(currentScrollOffset + 20, scrollRange + self.config.overflowHeight))
-    elseif delta > 0 and currentScrollOffset > self:GetHeight() then
-      self:SetVerticalScroll(currentScrollOffset - 20)
+    if delta < 0 then
+      -- Scroll down
+      scrollValue = math.min(self:GetVerticalScroll() + 20, maxScroll)
+    else
+      -- Scroll up
+      scrollValue = math.max(self:GetVerticalScroll() - 20, math.min(minScroll, maxScroll))
+    end
+
+    self:UpdateScrollChildRect()
+    self:SetVerticalScroll(scrollValue)
+
+    self.state.scrollAtBottom = scrollValue == maxScroll
+
+    -- Adjust height of scroll frame when scrolling
+    if self.state.scrollAtBottom then
+      -- If scrolled to the bottom, the height of the scroll frame should
+      -- include overflow to account for slide up animations
+      self:SetHeight(self.config.height + self.config.overflowHeight)
+      self.overlay:Hide()
+      self.overlay.newMessageHighlightFrame:Hide()
+      self.state.unreadMessages = false
+    else
+      -- If not, the height should fit the frame exactly so messages don't spill
+      -- under the edit box area
+      self:SetHeight(self.config.height)
+      self.overlay:Show()
     end
 
     -- Show hidden messages
     for _, message in ipairs(self.state.messages) do
-      if not message:IsVisible() then
-        message:Show()
-      end
+      message:Show()
     end
   end)
 
@@ -157,13 +327,13 @@ function SlidingMessageFrameMixin:Init(chatFrame)
     -- Don't hide chats when mouse is over
     self.state.mouseOver = true
 
-    for _, message in ipairs(self.state.messages) do
-      if Core.db.profile.chatShowOnMouseOver and not message:IsVisible() then
-        message:Show()
-      end
+    if not self.state.scrollAtBottom then
+      self.overlay:Show()
+    end
 
-      if message.outroTimer then
-        message.outroTimer:Cancel()
+    for _, message in ipairs(self.state.messages) do
+      if Core.db.profile.chatShowOnMouseOver then
+        message:Show()
       end
     end
   end)
@@ -172,14 +342,10 @@ function SlidingMessageFrameMixin:Init(chatFrame)
     -- Hide chats when mouse leaves
     self.state.mouseOver = false
 
+    self.overlay:HideDelay(Core.db.profile.chatHoldTime)
+
     for _, message in ipairs(self.state.messages) do
-      if message:IsVisible() then
-        message.outroTimer = C_Timer.NewTimer(Core.db.profile.chatHoldTime, function()
-          if message:IsVisible() then
-            message.outroAg:Play()
-          end
-        end)
-      end
+      message:HideDelay(Core.db.profile.chatHoldTime)
     end
   end)
 
@@ -192,10 +358,14 @@ function SlidingMessageFrameMixin:Init(chatFrame)
         self:SetHeight(self.config.height + self.config.overflowHeight)
         self:SetWidth(self.config.width)
 
-        self.slider:SetHeight(self.config.height + self.config.overflowHeight)
         self.slider:SetWidth(self.config.width)
 
-        self:SetVerticalScroll(self.slider:GetHeight() - self:GetHeight() + self.config.overflowHeight)
+        self.state.scrollAtBottom = true
+        self.state.unreadMessages = false
+        self:UpdateScrollChildRect()
+        self:SetVerticalScroll(self:GetVerticalScrollRange() + self.config.overflowHeight)
+        self.overlay:Hide()
+        self.overlay.newMessageHighlightFrame:Hide()
       end
 
       if key == "font" or key == "messageFontSize" or key == "frameWidth" or key == "frameHeight" then
@@ -244,7 +414,6 @@ function SlidingMessageFrameMixin:AddMessage(...)
   table.insert(self.state.incomingMessages, args)
 end
 
-
 function SlidingMessageFrameMixin:OnUpdate(elapsed)
   self.timeElapsed = self.timeElapsed + elapsed
   while (self.timeElapsed > 0.1) do
@@ -272,15 +441,26 @@ function SlidingMessageFrameMixin:Update()
     self.slider:SetHeight(newHeight)
     self.sliderTranslateUp:SetOffset(0, offset)
 
-    -- Display and run everything
-    self.sliderAg:SetScript("OnFinished", function ()
-      self:SetVerticalScroll(newHeight - self:GetHeight() + self.config.overflowHeight)
-    end)
-    self.sliderAg:Play()
+    -- Only play slide up if not scrolling
+    if self.state.scrollAtBottom then
+      -- Display and run everything
+      self.sliderAg:SetScript("OnFinished", function ()
+        self:SetVerticalScroll(newHeight - self:GetHeight() + self.config.overflowHeight)
+      end)
+      self.sliderAg:Play()
+    else
+      self.state.unreadMessages = true
+      self.overlay:Show()
+      self.overlay.newMessageHighlightFrame:Show()
+      self.overlay:HideDelay(Core.db.profile.chatHoldTime)
+    end
 
-    for _, messageFrame in ipairs(newMessages) do
-      messageFrame:Show()
-      table.insert(self.state.messages, messageFrame)
+    for _, message in ipairs(newMessages) do
+      message:Show()
+      if not self.state.mouseOver then
+        message:HideDelay(Core.db.profile.chatHoldTime)
+      end
+      table.insert(self.state.messages, message)
     end
 
     -- Release old messages